refactor: 코틀린 서버 폴더 분리
