+++ userserviceserver/src/main/kotlin/com/grepp/smartwatcha/app/api/controller/EmailVerification/EmailVerificationController.kt
@@ -1,10 +1,10 @@
-package com.grepp.smartwatcha.auth.controller
+package com.grepp.smartwatcha.app.api.controller.EmailVerification
 
-import com.grepp.smartwatcha.auth.dto.*
-import com.grepp.smartwatcha.auth.service.EmailVerificationService
+import com.grepp.smartwatcha.app.api.model.EmailVerification.dto.*
+import com.grepp.smartwatcha.app.api.model.EmailVerification.service.EmailVerificationService
+import jakarta.validation.Valid
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
-import javax.validation.Valid
 
 @RestController
 @RequestMapping("/api/v1/email-verification")
